swapped the old python for good old c++
