+++ requirments.txt
@@ -1,2 +1,3 @@
 yt-dlp
 crow
+curl
